Move MeshSettings1D and MeshSettings2D to Fragments and update references and namespaces
