+++ Lusas_Adapter/CRUD/Create/Properties/LineMesh.cs
@@ -22,6 +22,7 @@
 
 using Lusas.LPI;
 using BH.oM.Adapters.Lusas;
+using BH.oM.Adapters.Lusas.Fragments;
 using BH.oM.Structure.Elements;
 using BH.oM.Structure.Constraints;
 using BH.Engine.Adapter;
