+++ Lusas_Adapter/CRUD/Read/Properties/MeshSettings2D.cs
@@ -23,7 +23,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Lusas.LPI;
-using BH.oM.Adapters.Lusas;
+using BH.oM.Adapters.Lusas.Fragments;
 
 namespace BH.Adapter.Lusas
 {
