+++ Lusas_Adapter/Convert/ToBHoM/Elements/ToPanel.cs
@@ -22,6 +22,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
+using BH.oM.Adapters.Lusas.Fragments;
 using BH.oM.Structure.Elements;
 using BH.oM.Structure.SurfaceProperties;
 using BH.oM.Structure.Constraints;
