+++ Lusas_Adapter/Convert/ToBHoM/Properties/ToMeshSettings1D.cs
@@ -23,6 +23,7 @@
 using System.Linq;
 using Lusas.LPI;
 using BH.oM.Adapters.Lusas;
+using BH.oM.Adapters.Lusas.Fragments;
 using BH.Engine.Adapter;
 
 namespace BH.Adapter.Adapters.Lusas
