+++ Lusas_oM/Fragments/MeshSettings2D.cs
@@ -22,9 +22,9 @@
 
 using BH.oM.Base;
 
-namespace BH.oM.Adapters.Lusas
+namespace BH.oM.Adapters.Lusas.Fragments
 {
-    public class MeshSettings2D : BHoMObject
+    public class MeshSettings2D : BHoMObject, IFragment
     {
         /***************************************************/
         /****            Public Properties              ****/
